% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{capitalize.first}
\alias{capitalize.first}
\title{Capitalize the first letter of each word}
\usage{
capitalize.first(x)
}
\description{
Capitalize the first letter of each word
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
