% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleanup}
\alias{cleanup}
\title{Clean up the global environment to reduce memory use}
\usage{
cleanup(protect, threshold = 1)
}
\arguments{
\item{protect}{Vector of object names that user is not allowed to delete}

\item{threshold}{Objects with size below threshold are not offered for deletion. Megabytes. Defaults to 1.}
}
\description{
The function shows total memory use and lists all objects with a size larger than a threshold, giving the user the option to delete each one.
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
