% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{col.a}
\alias{col.a}
\title{Specify a transparent color with name and alpha value}
\usage{
col.a(colorname, alpha)
}
\arguments{
\item{colorname}{A color name recognized by col2rgb}

\item{alpha}{An alpha value between 0 and 1}
}
\description{
Specify a transparent color with name and alpha value
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
