% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colcheat}
\alias{colcheat}
\title{Plot all named colors}
\usage{
colcheat(cl = colors(), bg = "grey", cex = 0.75, rot = 30)
}
\description{
Plot all named colors
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
