% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{darjeeling}
\alias{darjeeling}
\title{A convenience function to use the wesanderson Darjeeling1 palette in a ggplot}
\usage{
darjeeling(ordering = T, ...)
}
\description{
A convenience function to use the wesanderson Darjeeling1 palette in a ggplot
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
