% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{Dobson.unit}
\alias{Dobson.unit}
\title{Constant to convert molec/cm2 to Dobson Units}
\format{An object of class \code{numeric} of length 1.}
\usage{
Dobson.unit
}
\description{
Constant to convert molec/cm2 to Dobson Units
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
\keyword{datasets}
