% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun.maxna}
\alias{fun.maxna}
\title{Calculate statistics with maximum amount of NA values specified}
\usage{
fun.maxna(x, fun, maxna)
}
\description{
A wrapper statistics functions that returns NA when there are more than maxna NAs in input
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
