% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geocode.loop}
\alias{geocode.loop}
\title{Geocode multiple locations, retrying in case of errors}
\usage{
geocode.loop(locs, repeats = 10, source = c("google", "dsk"), ...)
}
\arguments{
\item{locs}{Locations to geocode; a vector of place names or a data.frame or matrix with column called 'name' to be used as queries}

\item{repeats}{How many times to try before returning error}

\item{source}{See documentation for ggmap::geocode}
}
\value{
A data.frame with columns name, lat and lon
}
\description{
Geocode multiple locations, retrying in case of errors
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
