% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getadm}
\alias{getadm}
\title{Load administrative area boundaries}
\usage{
getadm(level = 0, path = Sys.getenv("GISpath"))
}
\arguments{
\item{level}{The level of admin boundaries to load.}

\item{path}{Path to the file}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Requires that the GADM RDS files or shapefiles exist in a path specified in either the environment variable GISpath or in the function parameter path. Tries to load first the RDS file and secondarily the shp file. If RDS file does not exist, it will convert the shp into RDS and write to disk.
}
\details{
The required files can be downloaded from https://gadm.org/
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
