% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getUTMproj}
\alias{getUTMproj}
\title{Determine the appropriate UTM proj4string to use for a location, or get proj4string for a zone}
\usage{
getUTMproj(zone = NULL, hem = NULL, loc = NULL, units = "km")
}
\arguments{
\item{zone}{Zone number; if missing, this is determined from loc}

\item{hem}{Hemisphere. "S" or "s" for southern. If missing, this is determined from loc.}

\item{loc}{Location. Either a Spatial* object or a vector with lon and lat values.}

\item{units}{Either "m" or "km"; defaults to km.}
}
\description{
Determine the appropriate UTM proj4string to use for a location, or get proj4string for a zone
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
