% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Return true for all values of x that are NOT included in y}
\usage{
x \%notin\% y
}
\description{
Return true for all values of x that are NOT included in y
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
