% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%whichin\%}
\alias{\%whichin\%}
\title{Return a vector of the values of x that are included in y}
\usage{
x \%whichin\% y
}
\description{
Return a vector of the values of x that are included in y
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
