% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%whichnotin\%}
\alias{\%whichnotin\%}
\title{Return a vector of the values of x that are NOT included in y}
\usage{
x \%whichnotin\% y
}
\description{
Return a vector of the values of x that are NOT included in y
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
