% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.outlier}
\alias{is.outlier}
\title{Simple outlier detection based on standard deviation}
\usage{
is.outlier(x, SDs = 10, na.rm = F)
}
\arguments{
\item{SDs}{How many standard deviations from mean signify an outlier. Defaults to 10.}
}
\description{
Simple outlier detection based on standard deviation
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
