% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{llproj}
\alias{llproj}
\title{Proj4string for standard latitude-longitude projection}
\format{An object of class \code{character} of length 1.}
\usage{
llproj
}
\description{
Proj4string for standard latitude-longitude projection
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
\keyword{datasets}
