% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mean.maxna}
\alias{mean.maxna}
\title{Calculate mean with maximum amount of NA values specified}
\usage{
\method{mean}{maxna}(x, maxna)
}
\description{
A wrapper around the base mean function that returns NA when there are more than maxna NAs in input
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
