% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{norm100}
\alias{norm100}
\title{Normalize data.frame column values to 100 in a specified column}
\usage{
norm100(df, cols = which(sapply(df, is.numeric)), basedate)
}
\arguments{
\item{cols}{Integer or character vector to indicate columns on which normalization should be performed. Defaults to all numeric columns.}

\item{basedate}{Integer or character to indicate column which is used as the basis for normalization. If several values are given, the second one is used in case value in the first one is not finite.}
}
\description{
Normalizes all values in data.frame columns based on values in one column. Most natural use is to set values to 100 on a basedate for plotting.
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
