% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{orderfactor}
\alias{orderfactor}
\title{Order factor levels based on another variable}
\usage{
orderfactor(var, by)
}
\arguments{
\item{var}{Factor or character vector with the values of the output factor.}

\item{by}{A vector that can be ordered with base::order.}
}
\description{
Create a factor from a factor or character string, with levels ordered based on another variable.
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
