% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{poscheat}
\alias{poscheat}
\title{Cheatsheet for lattice pos argument}
\usage{
poscheat()
}
\description{
Print a plot to show text label positions corresponding to different lattice pos arguments (1 to 4)
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
