% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quickpng}
\alias{quickpng}
\title{A convenience function to initialize a new png plot with defaults for width, height and res}
\usage{
quickpng(file, width = 2000, height = 1500, res = 300, ...)
}
\description{
A convenience function to initialize a new png plot with defaults for width, height and res
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
