% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sigfloor}
\alias{sigfloor}
\title{Round a number down to specified number of significant digits}
\usage{
sigfloor(x, sigdig = 1)
}
\description{
Round a number down to specified number of significant digits
}
\author{
Lauri Myllyvirta \email{lauri.myllyvirta@greenpeace.org}
}
